requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
paddleocr             # pulls paddlepaddle-runtime automatically
PyMuPDF==1.26.1       # explicit, ensures fitz module
pymupdf4llm>=0.0.23
opencv-python-headless
numpy
openai
python-dotenv
paddlepaddle-gpu==2.6.2